requirements.txt
aiohttp>=3.13.3
discord-py>=2.6.4
asyncpg>=0.29.0
